fix: 초단기 실황 날씨 api 수정 #S07P31A703-163

+++ backend/community/src/main/java/com/a703/community/dto/response/WeatherDto.java
@@ -9,7 +9,7 @@ public class WeatherDto {
 
     String weather;
 
-    Integer temperature;
+    Double temperature;
 
     String description;
 }
